# internal/server/openapi/swagger.yaml
basePath: /
consumes:
- application/json
definitions:
...
  handlers.createWorktreeRequest:
    properties:
      branch:
        type: string
      project:
        type: string
    type: object
  handlers.errorResponse:
    properties:
      message:
        type: string
    type: object
  handlers.projectResponse:
    properties:
      gitBranch:
        type: string
      githubUrl:
        type: string
      linearTicketUrl:
        type: string
      name:
        type: string
      pullRequestUrl:
        type: string
    type: object
...
  handlers.removeWorktreeRequest:
    properties:
      project:
        type: string
      worktree:
        type: string
    type: object
...
  handlers.worktreeResponse:
    properties:
      worktree:
        $ref: '#/definitions/worktree.Worktree'
    type: object
  handlers.worktreesResponse:
    properties:
      worktrees:
        items:
          $ref: '#/definitions/worktree.Worktree'
        type: array
    type: object
...
  worktree.RemoteBranch:
    properties:
      branch:
        type: string
      hasLocalBranch:
        type: boolean
      isCheckedOut:
        type: boolean
      name:
        type: string
      worktreeName:
        type: string
      worktreeProjectName:
        type: string
    type: object
  worktree.RemoteBranchList:
    properties:
      branches:
        items:
          $ref: '#/definitions/worktree.RemoteBranch'
        type: array
      remote:
        type: string
    type: object
  worktree.Worktree:
    properties:
      branch:
        type: string
      ignoredFileCopyWarnings:
        items:
          type: string
        type: array
      isBase:
        type: boolean
      isCurrent:
        type: boolean
      isRemovable:
        type: boolean
      name:
        type: string
      path:
        type: string
      projectName:
        type: string
    type: object
info:
  contact: {}
  description: Local HTTP API for the WADE browser workspace.
  title: WADE API
  version: 0.1.0
paths:
...
  /api/project:
    get:
      parameters:
      - description: Project name
        in: query
        name: project
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.projectResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.errorResponse'
      summary: Get project details
      tags:
      - Projects
...
  /api/terminal/reload:
    post:
      parameters:
      - description: Project name
        in: query
        name: project
        required: true
        type: string
      - description: Terminal name
        in: query
        name: terminal
        type: string
      - description: Agent name
        in: query
        name: agent
        type: string
      produces:
      - text/plain
      responses:
        "204":
          description: No Content
        "404":
          description: project not found
          schema:
            type: string
      summary: Reload terminal session
      tags:
      - Terminals
  /api/worktrees:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Worktree removal request
        in: body
        name: request
        schema:
          $ref: '#/definitions/handlers.removeWorktreeRequest'
      - description: Project name
        in: query
        name: project
        type: string
      - description: Worktree name
        in: query
        name: worktree
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.worktreeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.errorResponse'
      summary: Remove worktree
      tags:
      - Worktrees
    get:
      parameters:
      - description: Project name
        in: query
        name: project
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.worktreesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.errorResponse'
      summary: List worktrees
      tags:
      - Worktrees
    post:
      consumes:
      - application/json
      parameters:
      - description: Worktree creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.createWorktreeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.worktreeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.errorResponse'
      summary: Create worktree
      tags:
      - Worktrees
  /api/worktrees/remote-branches:
    get:
      parameters:
      - description: Project name
        in: query
        name: project
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/worktree.RemoteBranchList'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.errorResponse'
      summary: List remote branches
      tags:
      - Worktrees
produces:
- application/json
schemes:
- http
swagger: "2.0"